solving slow generation for large drawings
psupgrader cycles through every object before the filedialog opens,
takes a long time.   can we do without the upgrader?

in gcode generate detect that there is a selection and only try to
generate gcode for the selection.  this also saves a lot of time when
there is a lot of drawing outside the safearea.