removed shift from outline tool, added sticky size to large holes
pressing shift to move the cutline to inside./outside of a face is
outmoded and interferes with proper arc processing, removed.
having to hold shift and select a size for every large hole is now
avoided by having a lock mode so user can set a size and use it for all
future holes until reset or tool is reselected.  help updated.