plungecut throws exception when face is not phlat
when a face not on z-0 was clicked, you could create a hole that had
might not cut correctly because the face was not phlat, coloured line
appeared at an angle.
now detect z!=0 and excepts with user message

groupnames were long with high precision settings, add stripzeros to
trim the comment lengths.